setting up sharding using ActiveRecord::Base connections

+++ lib/tasks/regenerate_settings_from_local.rake
@@ -21,10 +21,16 @@
 namespace :data do
   desc "Regenerate settings table from local application.yml"
   task regenerate_settings: :environment do
-    #Truncate settings table
-    ActiveRecord::Base.connection.execute("TRUNCATE settings")
+    if ENV['DB']
+      ActiveRecord::Base.connected_to(database: ENV['DB'].to_sym) do
+        #Truncate settings table
+        ActiveRecord::Base.connection.execute("TRUNCATE settings")
 
-    #Rerun populator
-    SettingsPopulator.new().populate
+        #Rerun populator
+        SettingsPopulator.new().populate
+      end
+    else
+      puts "Please provide a database using \"DB=databasename\""
+    end
   end
 end
